Project's mdules refactor
